convert code to support 32 bit data

+++ HackPackageSource/Hack/Translators/HackTranslator.java
@@ -48,7 +48,7 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
     protected int programSize;
 
     // The program array
-    protected short[] program;
+    protected int[] program;
 
     // The source code array
     protected String[] source;
@@ -106,7 +106,7 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
      * If save is true, the compiled program will be saved automatically into a destination
      * file that will have the same name as the source but with the destination extension.
      */
-    public HackTranslator(String fileName, int size, short nullValue, boolean save)
+    public HackTranslator(String fileName, int size, int nullValue, boolean save)
      throws HackTranslatorException {
         if (fileName.indexOf(".") < 0)
             fileName = fileName + "." + getSourceExtension();
@@ -129,7 +129,7 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
      * A non null sourceFileName specifies a source file to be loaded.
      * The gui is assumed to be not null.
      */
-    public HackTranslator(HackTranslatorGUI gui, int size, short nullValue, String sourceFileName)
+    public HackTranslator(HackTranslatorGUI gui, int size, int nullValue, String sourceFileName)
      throws HackTranslatorException {
         this.gui = gui;
         gui.addHackTranslatorListener(this);
@@ -209,8 +209,8 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
     /**
      * initializes the HackTranslator.
      */
-    protected void init(int size, short nullValue) {
-        program = new short[size];
+    protected void init(int size, int nullValue) {
+        program = new int[size];
         for (int i = 0; i < size; i++)
             program[i] = nullValue;
         programSize = 0;
@@ -418,13 +418,13 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
      * If display is true, the version is for display purposes. Otherwise, the
      * version should be the final one.
      */
-    protected abstract String getCodeString(short code, int pc, boolean display);
+    protected abstract String getCodeString(int code, int pc, boolean display);
 
     /**
      * Adds the given command to the next position in the program.
      * Throws HackTranslatorException if the program is too large
      */
-    protected void addCommand(short command) throws HackTranslatorException {
+    protected void addCommand(int command) throws HackTranslatorException {
         if (destPC >= program.length)
             throw new HackTranslatorException("Program too large");
 
@@ -436,7 +436,7 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
     /**
      * Replaces the command in program location pc with the given command.
      */
-    protected void replaceCommand(int pc, short command) {
+    protected void replaceCommand(int pc, int command) {
         program[pc] = command;
         if (updateGUI)
             gui.getDestination().setLineAt(pc, getCodeString(command, pc, true));
@@ -578,7 +578,7 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
     /**
      * Returns the translated machine code program array
      */
-    public short[] getProgram() {
+    public int[] getProgram() {
         return program;
     }
 
@@ -586,7 +586,7 @@ public abstract class HackTranslator implements HackTranslatorEventListener, Act
      * Dumps the contents of the translated program into the destination file
      */
     private void dumpToFile() {
-        for (short i = 0; i < programSize; i++)
+        for (int i = 0; i < programSize; i++)
             writer.println(getCodeString(program[i], i, false));
         writer.close();
     }
